refactor: Split RateLimiterService into RateLimiterService and IndividualRateLimiter, RateLimiterService is now Injectable Service
